+++ examples/docs/src/components/docs/Sidebar.astro
@@ -1,5 +1,4 @@
 ---
-import { getCollection } from "astro:content";
 import SidebarTree from "./SidebarTree.astro";
 
 const { content, slugPath } = Astro.props;
