+++ examples/docs/tsconfig.json
@@ -7,8 +7,7 @@
     },
     "jsx": "react-jsx",
     "jsxImportSource": "react",
-    "module": "ES2022",
-    "types": ["vite/client"]
+    "module": "ES2022"
   },
   "include": [".astro/types.d.ts", "**/*"],
   "exclude": ["dist"]
