+++ packages/docs/README.md
@@ -1,24 +1,19 @@
-# unDocs
+# @korhq/undocs
 
-## v0.3.0-alpha
-
-Build your docs in an Astro site, using unstyled docs primitives and schemas with all the functionality built in. Style it with your own brand and layout. Integrate within your current website.
-
-This package provides schemas, recommended structures and helper functions to help you build your own docs UI. Right now, you have to build the UI and structure yourself. In the future, we'll have a CLI with presets available. We also will to offer headless, unstyled documentation components such as file trees, codeblocks, tabs, steps, and various plugins that you can style yourself.
+Build your docs in an Astro site, using unstyled docs primitives and schemas with all the functionality built in.
 
 ## Installation
 
-Make sure you have an Astro website ready to go, if you haven't made one already.
+```bash
+npm install @korhq/undocs
+```
 
-Install `@korhq/undocs` with your package manager.
+## Features
 
-## Docs Content Collection
+### Content Collections
 
-You can write docs with Markdown and MDX. Markdoc is untested right now.
+The package provides schemas and utilities for creating structured documentation using Astro's content collections.
 
-Make sure to use our schema from the `@korhq/undocs` package.
-
-**src/content.config.ts**
 ```ts
 import { defineCollection } from "astro:content";
 import { DocsSchema } from "@korhq/undocs";
@@ -32,141 +27,37 @@ const docs = defineCollection({
 export const collections = { docs };
 ```
 
-## Directories
+### React Component Previews
 
-You can organize your content in directories and subdirectories.
+Use the `reactComponentLoader` and `Playground` component to showcase React components in your documentation:
 
-In each directory, create an `index.json` file with `directory` set to true, and the directory title.
+```ts
+// content.config.ts
+import { defineCollection } from "astro:content";
+import { reactComponentLoader } from "@korhq/undocs";
+import { z } from "astro/zod";
 
-```json
+const previews = defineCollection({
+  schema: z.object({
+    title: z.string(),
+    component: z.string().optional(),
+  }).passthrough(),
+  loader: reactComponentLoader({ previewsDir: "src/previews" }),
+});
+
+export const collections = { docs, previews };
+```
+
+Then use the Playground component to render your React components:
+
+```astro
+---
+import { getCollection } from "astro:content";
+import { Playground } from "@korhq/undocs";
+
+const { preview } = Astro.props;
+const { component, title } = preview.data;
+---
+
+<h1>{title}</h1>
 {
-  "title": "Tesasdasdasdt",
-  "directory": true
-}
-```
-
-This will render directories (collapsibles) with the directory contents in the sidebar.
-
-## Components
-
-You should make your own static components such as Asides, Cards, etc.
-
-Right now we provide simple dynamic, headless, unstyled components made in Astro, such as:
-
-### Codeblock
-
-The codeblock primitive has a copy button that copies Astro's default markdown codeblock, wrapped in a container.
-
-Make your own `src/components/docs/Codeblock.astro`, importing from `@korhq/undocs/astro`.
-
-Here's an example with a wrapper, and a filename/langName.
-
-```astro
----
-import { Codeblock as CodeblockPrimitive } from "@korhq/undocs/astro";
-
-interface Props {
-  filename?: string;
-  langName: string;
-}
-const { filename, langName } = Astro.props;
----
-
-<CodeblockPrimitive
-  class="flex gap-1 p-2 flex-col h-fit text-white rounded-lg border border-white/10 mb-1 bg-neutral-950"
->
-  <div class="w-full flex justify-between">
-    <p class="text-xs font-mono">{filename || langName}</p>
-    <button class="text-xs active:text-orange-400" id="undocs-copy-button"> <!-- The copy button must have this ID -->
-      Copy
-    </button>
-  </div>
-  <div id="codeblock"><slot /></div> <!-- The slot must be wrapped in div#codeblock -->
-</CodeblockPrimitive>
-```
-
-And here's the usage of the example component:
-
-```mdx
-import Codeblock from "~/components/docs/Codeblock.astro";
-
-<Codeblock filename="goober.rs">
-  ```rust
-  fn main() {
-    println!("goober is probably the silliest kitty in existence");
-  }
-  ``
-</Codeblock>
-```
-
-### Tabs
-
-The tabs primitive tabs logic .
-
-Make your own `src/components/docs/Tabs.astro`, importing from `@korhq/undocs/astro`.
-
-Here's an example component using tailwind + css to style the buttons.
-
-```astro
----
-import { Tabs as TabsPrimitive } from "@korhq/undocs/astro";
----
-
-<TabsPrimitive class="flex flex-col gap-1 mb-1">
-  <div class="flex p-2 pb-0 rounded-xl rounded-b-sm bg-neutral-900"> <!-- Tab buttons container -->
-    <slot name="tab-buttons" />
-  </div>
-  <div class="bg-neutral-900 p-2 rounded-xl rounded-t-sm"> <!-- Tab contents container -->
-    <slot />
-  </div>
-</TabsPrimitive>
-
-<style>
-  /* Use tailwind to style the buttons: */
-  @reference "~/styles/global.css";
-
-  :global(button[data-tab-id]) { /* Tab buttons default state */
-    @apply px-2 border-b border-transparent;
-  }
-
-  :global(button[data-tab-id].active) { /* Tab buttons active state */
-    @apply border-b border-orange-400;
-  }
-
-  :global(div[data-tab-id]) {
-    @apply hidden;
-  }
-
-  :global(div[data-tab-id].active) {
-    @apply block;
-  }
-</style>
-
-```
-
-And here's the usage of the example component:
-
-```mdx
-import Tabs from '~/components/docs/Tabs.astro';
-
-<Tabs>
-  <Fragment slot="tab-buttons">
-    <button data-tab-id="1">
-      Tab 1
-    </button>
-    <button data-tab-id="2">
-      Tab 2
-    </button>
-  </Fragment>
-
-  <div data-tab-id="1">
-    Lorem ipsum
-  </div>
-  <div data-tab-id="2">
-    Dolar sit amet
-  </div>
-</Tabs>
-```
-The `Fragment` puts the buttons into the right slot, you can also apply the slot name to each button manually.
-
-The `data-tab-id` matches each button and contents.
