+++ packages/docs/package.json
@@ -25,9 +25,6 @@
   "files": [
     "./dist",
     "./src/components",
-    "./src/loaders",
-    "./src/schemas",
-    "./src/store",
     "LICENSE.txt"
   ],
   "dependencies": {
