+++ packages/docs/src/index.ts
@@ -1,6 +1,6 @@
-export * from "./schemas/Docs";
+export * from "./schemas/docs";
 export * from "./schemas/variant";
 export * from "./utilities/getCurrentLevelItems";
 
-export * from "./loaders/reactComponentLoader";
+export * from "./loaders/reactPreviewLoader";
 export * from "./store/variantsStore";
