+++ packages/docs/src/schemas/docs.ts
